Add files via upload
Updated the program, and now is somewhat working. Shows a window...a rather ugly window. In the next couple of days will add some more style to it...perhaps a cross in the background, or the Christian fish symbol. Also has only two Bible verses, but I will fix that tomorrow/later today. Whenever I find the time.